fix: solution order buttons visible in both themes
