+++ src/components/SwitchSolutionOrderButton/SwitchSolutionOrderButton.module.css
@@ -1,0 +1,31 @@
+.OrderButton {
+    border-bottom: solid 1px var(--text-primary);
+    border-right: solid 1px var(--text-primary);
+    border-left: none;
+    border-top: none;
+    background-color: transparent;
+    width: 20px;
+    height: 20px;
+
+    cursor: pointer;
+}
+
+.OrderButton:hover {
+    border-bottom-width: 2px;
+    border-right-width: 2px;
+}
+
+.OrderButtonUp {
+    transform: translateY(10%) rotate(225deg);
+}
+
+.OrderButtonDown {
+    transform: rotate(45deg) translateY(-80%);
+}
+
+.OrderButtonDisabled {
+    border-bottom-color: var(--text-ghost);
+    border-right-color: var(--text-ghost);
+
+    cursor: not-allowed;
+}
